requirements.txt
numpy==1.26.4
scipy==1.11.4
pandas==2.2.2
python-dotenv==1.0.0
pyodbc==5.1.0
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
prophet==1.1.5
openpyxl==3.1.2
